#321: move embed building out of `raiseToast()`
